Feat: add orca example (#149)
* feat: test orca custom

* feat: add get_desired_omni_vel

* feat: Add orca custom method

* del custom orca

* feat: Add orca example

* Delete custom_behavior_orca.py

* doc: add example

* test: improve coverage

* test: improve cov

* test: improve cov

* test: improve cov

+++ irsim/world/world.py
@@ -103,7 +103,7 @@ class World:
         Advance the simulation by one step.
         """
         self.count += 1
-        self.sampling = self.count % (self.sample_time / self.step_time) == 0
+        self.sampling = self.count % (int(self.sample_time / self.step_time)) == 0
 
         world_param.time = self.time
         world_param.count = self.count
